+++ proxypay/models.py
@@ -1,6 +1,6 @@
 import decimal
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from django.utils.timezone import now
 
 from django_admin_display import admin_display as d